Cleaned up boxes section in web manual and stb-boxes.less.
We will NOT have a box with several links inside it, like the contact box I just removed. That kind of need will be met by the use of lists of icons in the basic box example, or preferably without a box at all, like in the specialized contact links examples demonstrated in the Typography/Links section.